fix(vscode): handle ClinePass action fallback
