+++ apps/vscode/webview-ui/src/components/chat/chat-view/components/layout/WelcomeSection.tsx
@@ -19,24 +19,11 @@ import { useExtensionState } from "@/context/ExtensionStateContext"
 import { useHasFeatureFlag } from "@/hooks/useFeatureFlag"
 import { AccountServiceClient, StateServiceClient, UiServiceClient, WorktreeServiceClient } from "@/services/grpc-client"
 import { convertBannerData } from "@/utils/bannerUtils"
+import { buildClinePassSubscriptionUrl } from "@/utils/clinePassSubscription"
 import { getCurrentPlatform } from "@/utils/platformUtils"
 import { WelcomeSectionProps } from "../../types/chatTypes"
 
 const CLINE_PASS_PROMO_BANNER_ID = "cline-pass-home-promo-v2"
-const DEFAULT_CLINE_APP_BASE_URL = "https://app.cline.bot"
-const CLINE_PASS_SUBSCRIPTION_PATH = "dashboard/subscription"
-
-function buildClinePassSubscriptionUrl(appBaseUrl?: string): string {
-	try {
-		const baseUrl = appBaseUrl || DEFAULT_CLINE_APP_BASE_URL
-		const base = baseUrl.endsWith("/") ? baseUrl : `${baseUrl}/`
-		const url = new URL(CLINE_PASS_SUBSCRIPTION_PATH, base)
-		url.searchParams.set("personal", "true")
-		return url.toString()
-	} catch {
-		return `${DEFAULT_CLINE_APP_BASE_URL}/${CLINE_PASS_SUBSCRIPTION_PATH}?personal=true`
-	}
-}
 
 /**
  * Welcome section shown when there's no active task
